Reduce parser code
